requirements.txt
aniso8601==3.0.2
atomicwrites==1.2.1
attrs==18.2.0
click==6.7
Flask==1.0.2
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
marshmallow==3.0.0b13
marshmallow-sqlalchemy==0.14.1
more-itertools==4.3.0
numpy==1.15.1
pathlib2==2.3.2
Pillow==5.2.0
pkg-resources==0.0.0
pluggy==0.7.1
py==1.6.0
pytest==3.8.0
pytz==2018.5
six==1.11.0
SQLAlchemy==1.2.11
Werkzeug==0.14.1
